requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
atomicwrites=1.4.0=pyh9f0ad1d_0
attrs=19.3.0=py_0
blas=1.0=mkl
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py38h32f6830_0
colorama=0.4.3=py_0
cycler=0.10.0=py38_0
entrypoints=0.3=py38h32f6830_1001
flake8=3.7.9=py38h32f6830_1
flake8-polyfill=1.0.2=py_0
flake8-quotes=3.0.0=pyh9f0ad1d_0
freetype=2.9.1=ha9979f8_1
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=1.6.0=py38h32f6830_0
importlib_metadata=1.6.0=0
intel-openmp=2020.0=166
jpeg=9b=hb83a4c4_2
kiwisolver=1.0.1=py38ha925a31_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.3=py38_0
matplotlib-base=3.1.3=py38h64f37c6_0
mccabe=0.6.1=py_1
mkl=2020.0=166
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.0.15=py38h14836fe_0
mkl_random=1.1.0=py38hf9181ef_0
more-itertools=8.2.0=py_0
numpy=1.18.1=py38h93ca92e_0
numpy-base=1.18.1=py38hc3f5095_1
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1g=he774522_0
packaging=20.1=py_0
pep8-naming=0.10.0=pyh9f0ad1d_0
pillow=7.0.0=py38hcc1f983_0
pip=20.0.2=py38_1
pluggy=0.13.1=py38_0
py=1.8.1=py_0
py-opencv=4.0.1=py38he44ac1e_0
pycodestyle=2.5.0=py_0
pyflakes=2.1.1=py38_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py38ha925a31_4
pytesseract=0.3.4=pyh9f0ad1d_0
pytest=5.4.1=py38h32f6830_0
python=3.8.2=h5fd99cc_11
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
qt=5.9.7=vc14h73c81de_0
setuptools=46.1.3=py38_0
sip=4.19.13=py38ha925a31_0
six=1.14.0=py38_0
sqlite=3.31.1=h2a8f88b_1
tk=8.6.8=hfa6e2cd_0
tornado=6.0.4=py38he774522_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.9=pyh9f0ad1d_0
wheel=0.34.2=py38_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.3.7=h508b16e_0
